The first upload of "SpaceCol"
Realising that I should really be keeping a changelog of the work I'm
doing, I've upload SpaceCol to github. So far the vastly majority of
changes are cosmetic. The main changes at this point are changing the
green on the terrain to purple, increasing the yellow for plains, and
changing the names of the nations to companies, and giving them logos.